Rename test_cocotb.py to common and cleanup imports
