--- tests/test_cases/test_cocotb/test_cocotb.py
@@ -1,82 +1,0 @@
-#!/usr/bin/env python
-
-# Copyright (c) 2013, 2018 Potential Ventures Ltd
-# Copyright (c) 2013 SolarFlare Communications Inc
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#     * Redistributions of source code must retain the above copyright
-#       notice, this list of conditions and the following disclaimer.
-#     * Redistributions in binary form must reproduce the above copyright
-#       notice, this list of conditions and the following disclaimer in the
-#       documentation and/or other materials provided with the distribution.
-#     * Neither the name of Potential Ventures Ltd,
-#       SolarFlare Communications Inc nor the
-#       names of its contributors may be used to endorse or promote products
-#       derived from this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
-# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-# DISCLAIMED. IN NO EVENT SHALL POTENTIAL VENTURES LTD BE LIABLE FOR ANY
-# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-import logging
-import re
-import textwrap
-import traceback
-from fractions import Fraction
-from decimal import Decimal
-from math import isclose
-
-"""
-A set of tests that demonstrate cocotb functionality
-
-Also used as regression test of cocotb capabilities
-"""
-
-import cocotb
-from cocotb.triggers import (Timer, Join, RisingEdge, FallingEdge, Edge,
-                             ReadOnly, ReadWrite, ClockCycles, NextTimeStep,
-                             NullTrigger, Combine, Event, First, Trigger, Lock)
-from cocotb.clock import Clock
-from cocotb.result import TestFailure, TestError
-from cocotb.utils import get_sim_time
-from cocotb.outcomes import Value, Error
-
-
-@cocotb.coroutine
-def clock_gen(clock):
-    """Example clock gen for test use"""
-    for i in range(5):
-        clock <= 0
-        yield Timer(100)
-        clock <= 1
-        yield Timer(100)
-    clock._log.warning("Clock generator finished!")
-
-
-@cocotb.coroutine
-def _check_traceback(running_coro, exc_type, pattern):
-    try:
-        yield running_coro
-    except exc_type:
-        tb_text = traceback.format_exc()
-    else:
-        raise TestFailure("Exception was not raised")
-
-    if not re.match(pattern, tb_text):
-        raise TestFailure(
-            (
-                "Traceback didn't match - got:\n\n"
-                "{}\n"
-                "which did not match the pattern:\n\n"
-                "{}"
-            ).format(tb_text, pattern)
-        )
